frontend/requirements.txt
fastapi
uvicorn
streamlit
requests
pydantic
python-dotenv
pypdf
python-docx
